+++ utils/__init__.py
@@ -5,6 +5,7 @@ from .security import SecurityManager, security_manager
 from .cache_manager import CacheManager, cache_manager
 from .async_cache_manager import AsyncCacheManager, async_cache_manager
 from .git_manager import GitManager
+from .websocket_manager import WebSocketManager, websocket_manager, CollaborationSession
 from . import async_helpers
 
 __all__ = [
@@ -15,6 +16,9 @@ __all__ = [
     'AsyncCacheManager',
     'async_cache_manager',
     'GitManager',
+    'WebSocketManager',
+    'websocket_manager',
+    'CollaborationSession',
     'async_helpers'
 ]